app/requirements.txt
accelerate==1.6.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
audioread==3.0.1
azure-core==1.33.0
azure-storage-blob==12.25.1
bertopic==0.17.0
blis==0.7.11
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.16.0
colorama==0.4.6
comm==0.2.2
confection==0.1.5
cryptography==44.0.2
cymem==2.0.11
debugpy==1.8.13
decorator==5.2.1
dnspython==2.7.0
email_validator==2.2.0
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl#sha256=293e9547a655b25499198ab15a525b05b9407a75f10255e405e8c3854329ab63
executing==2.2.0
faiss-cpu==1.10.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
fsspec==2025.3.2
gensim==4.3.3
h11==0.14.0
hdbscan==0.8.40
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keybert==0.9.0
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.6.0
moviepy==2.1.2
mpmath==1.3.0
msgpack==1.1.0
murmurhash==1.0.12
narwhals==1.34.1
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numba==0.61.0
numpy==1.26.4
openai-whisper==20240930
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.7
plotly==6.0.1
pooch==1.8.2
preshed==3.0.9
proglog==0.1.11
prompt_toolkit==3.0.50
psutil==7.0.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
pynndescent==0.5.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==4.0.2
sentencepiece==0.2.0
setuptools==78.1.0
shellingham==1.5.4
six==1.17.0
smart-open==6.4.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
spacy==3.8.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
starlette==0.46.2
sympy==1.13.1
thinc==8.2.5
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.0
typer==0.9.4
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
umap-learn==0.5.7
urllib3==2.3.0
uvicorn==0.34.1
wasabi==1.1.3
watchfiles==1.0.5
wcwidth==0.2.13
weasel==0.3.4
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.2
